dagster_sap_loc/dagster_sap_tests/test_assets/test_sap_etl_dwh.py

import unittest
from unittest.mock import create_autospec
from dagster_sap.assets.sap_etl_dwh import DL_SAP_T001
from dagster import AssetExecutionContext, build_asset_context
from dagster_shared_gf.resources.sql_server_resources import SQLServerResource

class TestDL_SAP_T001(unittest.TestCase):
    def setUp(self):
        # Create a mock AssetExecutionContext
        self.context = build_asset_context()
        
        # Create a mock SQLServerResource
        self.mock_sql_server_resource = create_autospec(SQLServerResource)
        
        # Set up mock return values and behavior
        # Example: Configure the SQLServerResource mock
        self.mock_sql_server_resource.query.return_value = [
            {'column1': 'value1', 'column2': 'value2'},
            {'column1': 'value3', 'column2': 'value4'}
        ]
    
    def test_DL_SAP_T001(self):
        # Call the function with mocks
        result = DL_SAP_T001(context=self.context, dwh_farinter_dl=self.mock_sql_server_resource)
        
        # Assert expected behaviors and results
        self.mock_sql_server_resource.query.assert_called_once_with("expected_query_here")
        
        # Add more assertions based on the function's expected behavior
        # Example: Check if result has the expected structure or values
        self.assertIsNotNone(result)
        self.assertEqual(len(result), 2)
        self.assertEqual(result[0]['column1'], 'value1')

if __name__ == '__main__':
    unittest.main()
